--- src/Microsoft.AspNet.Server.Kestrel/Infrastructure/LongExtensions.cs
@@ -1,24 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNet.Server.Kestrel.Extensions
-{
-    public static class LongExtensions
-    {
-        public static int BitCount(this long value)
-        {
-            // see https://github.com/dotnet/corefx/blob/master/src/System.Reflection.Metadata/src/System/Reflection/Internal/Utilities/BitArithmetic.cs
-
-            const ulong Mask01010101 = 0x5555555555555555UL;
-            const ulong Mask00110011 = 0x3333333333333333UL;
-            const ulong Mask00001111 = 0x0F0F0F0F0F0F0F0FUL;
-            const ulong Mask00000001 = 0x0101010101010101UL;
-
-            var v = (ulong)value;
-
-            v = v - ((v >> 1) & Mask01010101);
-            v = (v & Mask00110011) + ((v >> 2) & Mask00110011);
-            return (int)(unchecked(((v + (v >> 4)) & Mask00001111) * Mask00000001) >> 56);
-        }
-    }
-}
